Takes a folder of Luciernaga_QC signatures and returns a data.frame of the signatures contained within, for use in downstream data analysis.

Takes a folder of Luciernaga_QC signatures and returns a data.frame of the signatures contained within, for use in downstream data analysis.

Usage

Luciernaga_FolderSignatures(
  FolderPath,
  sample.name,
  StringRemoval = NULL,
  fluorophore.name,
  Verbose = FALSE,
  stats = "median",
  PanelCuts = NULL,
  normalize = TRUE
)

Arguments

FolderPath

Location of the Luciernaga_QC .fcs outputs

sample.name

The keyword where the identifying sample name can be found

StringRemoval

Default NULL, provide to remove items from sample.name based on values found on the keyword

fluorophore.name

Specify the name of the fluorophore

Verbose

Default FALSE, provides info as it goes

stats

Whether to use median or mean

PanelCuts

Default NULL, provide a c(0.5,1) argument to specify the brightness percentiles to retrieve signature from for the individual files

normalize

Default TRUE, whether to return normalized or raw averaged MFI signatures

Value

A data.frame of the corresponding averaged signatures per file

Examples

A <- 2 + 2