Iterates over single color variant .fcs files, passing the generated signature to unmix full-stain sample

Iterates over single color variant .fcs files, passing the generated signature to unmix full-stain sample

Usage

Luciernaga_IterativeUnmixing(
  IterativePath,
  IterativeSampleName,
  Iterativeremovestrings,
  fluorophore.name,
  PanelCuts = NULL,
  stats = "median",
  samplecolumn = "Sample",
  controlData,
  FullStainedGS,
  sample.name,
  removestrings,
  subset = "root",
  returnType = "fcs",
  outpath = outpath,
  PanelPath
)

Arguments

IterativePath

File path to folder containing the variant single color .fcs files

PanelCuts

A data.frame containing the cuts, see Luciernaga_Unmixing.

stats

Whether to use "median" or "mean" to derrive average signatures

controlData

The Luciernaga_SingleColors output that will be used in combination with the iterated variant

FullStainedGS

The GatingSet containing the full-stained samples you want to see effect of unmixing on

sample.name

Keyword storage name

removestrings

Regular string values to remove to match single color name in controlData

subset

What population to pass from GatingSet, default is root.

outpath

Desired storage location

PanelPath

Path to a reference panel for ordering of column markers in unmixed file

iterate_removestrings

String values to remove from the variant .fcs files name

Verbose

If TRUE returns steps

SignatureView

Whether to visualize the SCs

returntype

Default is set to "fcs" for iteration arguments

Value

Iterated versions of the full-stained sample

Examples

NULL
#> NULL