Taking a Luciernaga_Slices (or similar data.frame) with an added Counts column, expands each row by the count, takes the median, and then returns either data or a plot showing how different the average signature is from the individual ones

Taking a Luciernaga_Slices (or similar data.frame) with an added Counts column, expands each row by the count, takes the median, and then returns either data or a plot showing how different the average signature is from the individual ones

Usage

QC_Amalgamate(
  data,
  samplecolumn = "Percentiles",
  normalize = FALSE,
  countcolumn = "Count",
  returnType = "plot",
  titlename = NULL,
  linecolor = "red",
  legend = TRUE
)

Arguments

data

The modified data.frame containing the sample, count and all the normalized detector columns

samplecolumn

The column name correspondoing to the sample column, default is Percentiles

countcolumn

The column name corresponding to the count column, default Count

returnType

Whether to return plot (default) or data

titlename

Default is NULL, specify a title if desired

linecolor

Default is red, specify alternate color for the Average signature

legend

Whether to return legend, default is TRUE

Value

Either a plot or the underlying data

Examples

A <- 2+2