Parses Library Reference Control .XML files and returns

Parses Library Reference Control .XML files and returns

Usage

QC_LibraryParse(
  x,
  returntype,
  references = TRUE,
  myfactor = "Fluorophore",
  namefactor = "Sample"
)

Arguments

x

An .XML file

returntype

What to return "data" or "plots"

references

Plot argument, adds red reference signature

myfactor

Plot argument, data column to group by for plotting. Default "Fluorophore".

namefactor

Plot argument, data column name added to Plot Title.

Value

TBD

Examples

Folder_Location <- system.file("extdata", package = "Luciernaga")
XML_Pattern <- ".XML$"
XML_Files <- list.files(path = Folder_Location, pattern = XML_Pattern,
                        full.names = TRUE, recursive = FALSE)
SinglePlot <- QC_LibraryParse(XML_Files[2], returntype="plots", references=FALSE)