Takes the dataframe generated from CytekQCFilePrep and returns all the QC plots

Takes the dataframe generated from CytekQCFilePrep and returns all the QC plots

Usage

QC_Plots(
  x,
  FailedFlag,
  MeasurementType = NULL,
  Metadata = NULL,
  plotType = "individual",
  returntype,
  path,
  filename,
  thecolumns = 1,
  therows = 3,
  width = 7,
  height = 9
)

Arguments

x

The data frame object of Aurora QC data generated by CytekQCFilePrep

FailedFlag

Whether to show red flags when a detectors "Out-of-Range" is TRUE

MeasurementType

A list containing character values, will select any present from the column names for plotting

Metadata

An optional character column to use as ggplot factor.

plotType

Whether to return "individual" plots or "comparison" plots (provide Metadata name!)

returntype

Whether to return to pdf, patchwork or plots.

path

Location path to return the pdf

filename

File name to save as.

thecolumns

The number of columns per page

therows

The number of rows per page

width

Desired page width

height

Desired page height

Value

The pdf and/the plots.

Examples

library(ggplot2)

File_Location <- system.file("extdata", package = "Luciernaga")
CSV_Pattern <- ".CSV$"
CSV_Files <- list.files(path=File_Location, pattern=CSV_Pattern,
                       full.names=TRUE)
TidyData <- QC_FilePrep(CSV_Files, TrackChange = FALSE)
#> Lines extracted from 2 to 227 written to LineChunk_1.txt 
#> Lines extracted from 230 to 455 written to LineChunk_2.txt 
#> Lines extracted from 458 to 683 written to LineChunk_3.txt 
#> Lines extracted from 686 to 911 written to LineChunk_4.txt 
#> Lines extracted from 914 to 1139 written to LineChunk_5.txt 
#> New names:
#> • `DateTime` -> `DateTime...1`
#> • `Out of Range Flag` -> `Out of Range Flag...3`
#> • `Out of Range Flag` -> `Out of Range Flag...5`
#> • `Out of Range Flag` -> `Out of Range Flag...7`
#> • `Out of Range Flag` -> `Out of Range Flag...9`
#> • `Out of Range Flag` -> `Out of Range Flag...11`
#> • `Out of Range Flag` -> `Out of Range Flag...13`
#> • `Out of Range Flag` -> `Out of Range Flag...15`
#> • `Out of Range Flag` -> `Out of Range Flag...17`
#> • `Out of Range Flag` -> `Out of Range Flag...19`
#> • `Out of Range Flag` -> `Out of Range Flag...21`
#> • `Out of Range Flag` -> `Out of Range Flag...23`
#> • `Out of Range Flag` -> `Out of Range Flag...25`
#> • `Out of Range Flag` -> `Out of Range Flag...27`
#> • `Out of Range Flag` -> `Out of Range Flag...29`
#> • `Out of Range Flag` -> `Out of Range Flag...31`
#> • `Out of Range Flag` -> `Out of Range Flag...33`
#> • `Out of Range Flag` -> `Out of Range Flag...35`
#> • `Out of Range Flag` -> `Out of Range Flag...37`
#> • `Out of Range Flag` -> `Out of Range Flag...39`
#> • `Out of Range Flag` -> `Out of Range Flag...41`
#> • `Out of Range Flag` -> `Out of Range Flag...43`
#> • `Out of Range Flag` -> `Out of Range Flag...45`
#> • `Out of Range Flag` -> `Out of Range Flag...47`
#> • `Out of Range Flag` -> `Out of Range Flag...49`
#> • `Out of Range Flag` -> `Out of Range Flag...51`
#> • `Out of Range Flag` -> `Out of Range Flag...53`
#> • `Out of Range Flag` -> `Out of Range Flag...55`
#> • `Out of Range Flag` -> `Out of Range Flag...57`
#> • `Out of Range Flag` -> `Out of Range Flag...59`
#> • `Out of Range Flag` -> `Out of Range Flag...61`
#> • `Out of Range Flag` -> `Out of Range Flag...63`
#> • `Out of Range Flag` -> `Out of Range Flag...65`
#> • `Out of Range Flag` -> `Out of Range Flag...67`
#> • `Out of Range Flag` -> `Out of Range Flag...69`
#> • `Out of Range Flag` -> `Out of Range Flag...71`
#> • `Out of Range Flag` -> `Out of Range Flag...73`
#> • `Out of Range Flag` -> `Out of Range Flag...75`
#> • `Out of Range Flag` -> `Out of Range Flag...77`
#> • `Out of Range Flag` -> `Out of Range Flag...79`
#> • `Out of Range Flag` -> `Out of Range Flag...81`
#> • `Out of Range Flag` -> `Out of Range Flag...83`
#> • `Out of Range Flag` -> `Out of Range Flag...85`
#> • `Out of Range Flag` -> `Out of Range Flag...87`
#> • `Out of Range Flag` -> `Out of Range Flag...89`
#> • `Out of Range Flag` -> `Out of Range Flag...91`
#> • `Out of Range Flag` -> `Out of Range Flag...93`
#> • `Out of Range Flag` -> `Out of Range Flag...95`
#> • `Out of Range Flag` -> `Out of Range Flag...97`
#> • `Out of Range Flag` -> `Out of Range Flag...99`
#> • `Out of Range Flag` -> `Out of Range Flag...101`
#> • `Out of Range Flag` -> `Out of Range Flag...103`
#> • `Out of Range Flag` -> `Out of Range Flag...105`
#> • `Out of Range Flag` -> `Out of Range Flag...107`
#> • `Out of Range Flag` -> `Out of Range Flag...109`
#> • `Out of Range Flag` -> `Out of Range Flag...111`
#> • `Out of Range Flag` -> `Out of Range Flag...113`
#> • `Out of Range Flag` -> `Out of Range Flag...115`
#> • `Out of Range Flag` -> `Out of Range Flag...117`
#> • `Out of Range Flag` -> `Out of Range Flag...119`
#> • `Out of Range Flag` -> `Out of Range Flag...121`
#> • `Out of Range Flag` -> `Out of Range Flag...123`
#> • `Out of Range Flag` -> `Out of Range Flag...125`
#> • `Out of Range Flag` -> `Out of Range Flag...127`
#> • `Out of Range Flag` -> `Out of Range Flag...129`
#> • `Out of Range Flag` -> `Out of Range Flag...131`
#> • `Out of Range Flag` -> `Out of Range Flag...133`
#> • `Out of Range Flag` -> `Out of Range Flag...135`
#> • `DateTime` -> `DateTime...136`
#> • `Out of Range Flag` -> `Out of Range Flag...138`
#> • `Out of Range Flag` -> `Out of Range Flag...140`
#> • `Out of Range Flag` -> `Out of Range Flag...142`
#> • `Out of Range Flag` -> `Out of Range Flag...144`
#> • `Out of Range Flag` -> `Out of Range Flag...146`
#> • `Out of Range Flag` -> `Out of Range Flag...148`
#> • `Out of Range Flag` -> `Out of Range Flag...150`
#> • `Out of Range Flag` -> `Out of Range Flag...152`
#> • `Out of Range Flag` -> `Out of Range Flag...154`
#> • `Out of Range Flag` -> `Out of Range Flag...156`
#> • `Out of Range Flag` -> `Out of Range Flag...158`
#> • `Out of Range Flag` -> `Out of Range Flag...160`
#> • `Out of Range Flag` -> `Out of Range Flag...162`
#> • `Out of Range Flag` -> `Out of Range Flag...164`
#> • `Out of Range Flag` -> `Out of Range Flag...166`
#> • `Out of Range Flag` -> `Out of Range Flag...168`
#> • `Out of Range Flag` -> `Out of Range Flag...170`
#> • `Out of Range Flag` -> `Out of Range Flag...172`
#> • `Out of Range Flag` -> `Out of Range Flag...174`
#> • `Out of Range Flag` -> `Out of Range Flag...176`
#> • `Out of Range Flag` -> `Out of Range Flag...178`
#> • `Out of Range Flag` -> `Out of Range Flag...180`
#> • `Out of Range Flag` -> `Out of Range Flag...182`
#> • `Out of Range Flag` -> `Out of Range Flag...184`
#> • `Out of Range Flag` -> `Out of Range Flag...186`
#> • `Out of Range Flag` -> `Out of Range Flag...188`
#> • `Out of Range Flag` -> `Out of Range Flag...190`
#> • `Out of Range Flag` -> `Out of Range Flag...192`
#> • `Out of Range Flag` -> `Out of Range Flag...194`
#> • `Out of Range Flag` -> `Out of Range Flag...196`
#> • `Out of Range Flag` -> `Out of Range Flag...198`
#> • `Out of Range Flag` -> `Out of Range Flag...200`
#> • `Out of Range Flag` -> `Out of Range Flag...202`
#> • `Out of Range Flag` -> `Out of Range Flag...204`
#> • `Out of Range Flag` -> `Out of Range Flag...206`
#> • `Out of Range Flag` -> `Out of Range Flag...208`
#> • `Out of Range Flag` -> `Out of Range Flag...210`
#> • `Out of Range Flag` -> `Out of Range Flag...212`
#> • `Out of Range Flag` -> `Out of Range Flag...214`
#> • `Out of Range Flag` -> `Out of Range Flag...216`
#> • `Out of Range Flag` -> `Out of Range Flag...218`
#> • `Out of Range Flag` -> `Out of Range Flag...220`
#> • `Out of Range Flag` -> `Out of Range Flag...222`
#> • `Out of Range Flag` -> `Out of Range Flag...224`
#> • `Out of Range Flag` -> `Out of Range Flag...226`
#> • `Out of Range Flag` -> `Out of Range Flag...228`
#> • `Out of Range Flag` -> `Out of Range Flag...230`
#> • `Out of Range Flag` -> `Out of Range Flag...232`
#> • `Out of Range Flag` -> `Out of Range Flag...234`
#> • `Out of Range Flag` -> `Out of Range Flag...236`
#> • `Out of Range Flag` -> `Out of Range Flag...238`
#> • `Out of Range Flag` -> `Out of Range Flag...240`
#> • `Out of Range Flag` -> `Out of Range Flag...242`
#> • `Out of Range Flag` -> `Out of Range Flag...244`
#> • `Out of Range Flag` -> `Out of Range Flag...246`
#> • `Out of Range Flag` -> `Out of Range Flag...248`
#> • `Out of Range Flag` -> `Out of Range Flag...250`
#> • `Out of Range Flag` -> `Out of Range Flag...252`
#> • `Out of Range Flag` -> `Out of Range Flag...254`
#> • `Out of Range Flag` -> `Out of Range Flag...256`
#> • `Out of Range Flag` -> `Out of Range Flag...258`
#> • `Out of Range Flag` -> `Out of Range Flag...260`
#> • `Out of Range Flag` -> `Out of Range Flag...262`
#> • `Out of Range Flag` -> `Out of Range Flag...264`
#> • `Out of Range Flag` -> `Out of Range Flag...266`
#> • `Out of Range Flag` -> `Out of Range Flag...268`
#> • `Out of Range Flag` -> `Out of Range Flag...270`
#> • `DateTime` -> `DateTime...271`
#> • `Out of Range Flag` -> `Out of Range Flag...273`
#> • `Out of Range Flag` -> `Out of Range Flag...275`
#> • `Out of Range Flag` -> `Out of Range Flag...277`
#> • `Out of Range Flag` -> `Out of Range Flag...279`
#> • `` -> `...280`
#> • `` -> `...281`
#> • `` -> `...282`
#> • `` -> `...283`
#> • `` -> `...284`
#> • `` -> `...285`
#> • `` -> `...286`
#> • `` -> `...287`
#> • `` -> `...288`
#> • `` -> `...289`
#> • `` -> `...290`
#> • `` -> `...291`
#> • `` -> `...292`
#> • `` -> `...293`
#> • `` -> `...294`
#> • `` -> `...295`
#> • `` -> `...296`
#> • `` -> `...297`
#> • `` -> `...298`
#> • `` -> `...299`
#> • `` -> `...300`
#> • `` -> `...301`
#> • `` -> `...302`
#> • `` -> `...303`
#> • `` -> `...304`
#> • `` -> `...305`
#> • `` -> `...306`
#> • `` -> `...307`
#> • `` -> `...308`
#> • `` -> `...309`
#> • `` -> `...310`
#> • `` -> `...311`
#> • `` -> `...312`
#> • `` -> `...313`
#> • `` -> `...314`
#> • `` -> `...315`
#> • `` -> `...316`
#> • `` -> `...317`
#> • `` -> `...318`
#> • `` -> `...319`
#> • `` -> `...320`
#> • `` -> `...321`
#> • `` -> `...322`
#> • `` -> `...323`
#> • `` -> `...324`
#> • `` -> `...325`
#> • `` -> `...326`
#> • `` -> `...327`
#> • `` -> `...328`
#> • `` -> `...329`
#> • `` -> `...330`
#> • `` -> `...331`
#> • `` -> `...332`
#> • `` -> `...333`
#> • `` -> `...334`
#> • `` -> `...335`
#> • `` -> `...336`
#> • `` -> `...337`
#> • `` -> `...338`
#> • `` -> `...339`
#> • `` -> `...340`
#> • `` -> `...341`
#> • `` -> `...342`
#> • `` -> `...343`
#> • `` -> `...344`
#> • `` -> `...345`
#> • `` -> `...346`
#> • `` -> `...347`
#> • `` -> `...348`
#> • `` -> `...349`
#> • `` -> `...350`
#> • `` -> `...351`
#> • `` -> `...352`
#> • `` -> `...353`
#> • `` -> `...354`
#> • `` -> `...355`
#> • `` -> `...356`
#> • `` -> `...357`
#> • `` -> `...358`
#> • `` -> `...359`
#> • `` -> `...360`
#> • `` -> `...361`
#> • `` -> `...362`
#> • `` -> `...363`
#> • `` -> `...364`
#> • `` -> `...365`
#> • `` -> `...366`
#> • `` -> `...367`
#> • `` -> `...368`
#> • `` -> `...369`
#> • `` -> `...370`
#> • `` -> `...371`
#> • `` -> `...372`
#> • `` -> `...373`
#> • `` -> `...374`
#> • `` -> `...375`
#> • `` -> `...376`
#> • `` -> `...377`
#> • `` -> `...378`
#> • `` -> `...379`
#> • `` -> `...380`
#> • `` -> `...381`
#> • `` -> `...382`
#> • `` -> `...383`
#> • `` -> `...384`
#> • `` -> `...385`
#> • `` -> `...386`
#> • `` -> `...387`
#> • `` -> `...388`
#> • `` -> `...389`
#> • `` -> `...390`
#> • `` -> `...391`
#> • `` -> `...392`
#> • `` -> `...393`
#> • `` -> `...394`
#> • `` -> `...395`
#> • `` -> `...396`
#> • `` -> `...397`
#> • `` -> `...398`
#> • `` -> `...399`
#> • `` -> `...400`
#> • `` -> `...401`
#> • `` -> `...402`
#> • `` -> `...403`
#> • `` -> `...404`
#> • `` -> `...405`
#> • `DateTime` -> `DateTime...406`
#> • `Out of Range Flag` -> `Out of Range Flag...408`
#> • `Out of Range Flag` -> `Out of Range Flag...410`
#> • `Out of Range Flag` -> `Out of Range Flag...412`
#> • `Out of Range Flag` -> `Out of Range Flag...414`
#> • `Out of Range Flag` -> `Out of Range Flag...416`
#> • `` -> `...417`
#> • `` -> `...418`
#> • `` -> `...419`
#> • `` -> `...420`
#> • `` -> `...421`
#> • `` -> `...422`
#> • `` -> `...423`
#> • `` -> `...424`
#> • `` -> `...425`
#> • `` -> `...426`
#> • `` -> `...427`
#> • `` -> `...428`
#> • `` -> `...429`
#> • `` -> `...430`
#> • `` -> `...431`
#> • `` -> `...432`
#> • `` -> `...433`
#> • `` -> `...434`
#> • `` -> `...435`
#> • `` -> `...436`
#> • `` -> `...437`
#> • `` -> `...438`
#> • `` -> `...439`
#> • `` -> `...440`
#> • `` -> `...441`
#> • `` -> `...442`
#> • `` -> `...443`
#> • `` -> `...444`
#> • `` -> `...445`
#> • `` -> `...446`
#> • `` -> `...447`
#> • `` -> `...448`
#> • `` -> `...449`
#> • `` -> `...450`
#> • `` -> `...451`
#> • `` -> `...452`
#> • `` -> `...453`
#> • `` -> `...454`
#> • `` -> `...455`
#> • `` -> `...456`
#> • `` -> `...457`
#> • `` -> `...458`
#> • `` -> `...459`
#> • `` -> `...460`
#> • `` -> `...461`
#> • `` -> `...462`
#> • `` -> `...463`
#> • `` -> `...464`
#> • `` -> `...465`
#> • `` -> `...466`
#> • `` -> `...467`
#> • `` -> `...468`
#> • `` -> `...469`
#> • `` -> `...470`
#> • `` -> `...471`
#> • `` -> `...472`
#> • `` -> `...473`
#> • `` -> `...474`
#> • `` -> `...475`
#> • `` -> `...476`
#> • `` -> `...477`
#> • `` -> `...478`
#> • `` -> `...479`
#> • `` -> `...480`
#> • `` -> `...481`
#> • `` -> `...482`
#> • `` -> `...483`
#> • `` -> `...484`
#> • `` -> `...485`
#> • `` -> `...486`
#> • `` -> `...487`
#> • `` -> `...488`
#> • `` -> `...489`
#> • `` -> `...490`
#> • `` -> `...491`
#> • `` -> `...492`
#> • `` -> `...493`
#> • `` -> `...494`
#> • `` -> `...495`
#> • `` -> `...496`
#> • `` -> `...497`
#> • `` -> `...498`
#> • `` -> `...499`
#> • `` -> `...500`
#> • `` -> `...501`
#> • `` -> `...502`
#> • `` -> `...503`
#> • `` -> `...504`
#> • `` -> `...505`
#> • `` -> `...506`
#> • `` -> `...507`
#> • `` -> `...508`
#> • `` -> `...509`
#> • `` -> `...510`
#> • `` -> `...511`
#> • `` -> `...512`
#> • `` -> `...513`
#> • `` -> `...514`
#> • `` -> `...515`
#> • `` -> `...516`
#> • `` -> `...517`
#> • `` -> `...518`
#> • `` -> `...519`
#> • `` -> `...520`
#> • `` -> `...521`
#> • `` -> `...522`
#> • `` -> `...523`
#> • `` -> `...524`
#> • `` -> `...525`
#> • `` -> `...526`
#> • `` -> `...527`
#> • `` -> `...528`
#> • `` -> `...529`
#> • `` -> `...530`
#> • `` -> `...531`
#> • `` -> `...532`
#> • `` -> `...533`
#> • `` -> `...534`
#> • `` -> `...535`
#> • `` -> `...536`
#> • `` -> `...537`
#> • `` -> `...538`
#> • `` -> `...539`
#> • `` -> `...540`
#> • `DateTime` -> `DateTime...541`
#> • `Out of Range Flag` -> `Out of Range Flag...543`
#> • `` -> `...544`
#> • `` -> `...545`
#> • `` -> `...546`
#> • `` -> `...547`
#> • `` -> `...548`
#> • `` -> `...549`
#> • `` -> `...550`
#> • `` -> `...551`
#> • `` -> `...552`
#> • `` -> `...553`
#> • `` -> `...554`
#> • `` -> `...555`
#> • `` -> `...556`
#> • `` -> `...557`
#> • `` -> `...558`
#> • `` -> `...559`
#> • `` -> `...560`
#> • `` -> `...561`
#> • `` -> `...562`
#> • `` -> `...563`
#> • `` -> `...564`
#> • `` -> `...565`
#> • `` -> `...566`
#> • `` -> `...567`
#> • `` -> `...568`
#> • `` -> `...569`
#> • `` -> `...570`
#> • `` -> `...571`
#> • `` -> `...572`
#> • `` -> `...573`
#> • `` -> `...574`
#> • `` -> `...575`
#> • `` -> `...576`
#> • `` -> `...577`
#> • `` -> `...578`
#> • `` -> `...579`
#> • `` -> `...580`
#> • `` -> `...581`
#> • `` -> `...582`
#> • `` -> `...583`
#> • `` -> `...584`
#> • `` -> `...585`
#> • `` -> `...586`
#> • `` -> `...587`
#> • `` -> `...588`
#> • `` -> `...589`
#> • `` -> `...590`
#> • `` -> `...591`
#> • `` -> `...592`
#> • `` -> `...593`
#> • `` -> `...594`
#> • `` -> `...595`
#> • `` -> `...596`
#> • `` -> `...597`
#> • `` -> `...598`
#> • `` -> `...599`
#> • `` -> `...600`
#> • `` -> `...601`
#> • `` -> `...602`
#> • `` -> `...603`
#> • `` -> `...604`
#> • `` -> `...605`
#> • `` -> `...606`
#> • `` -> `...607`
#> • `` -> `...608`
#> • `` -> `...609`
#> • `` -> `...610`
#> • `` -> `...611`
#> • `` -> `...612`
#> • `` -> `...613`
#> • `` -> `...614`
#> • `` -> `...615`
#> • `` -> `...616`
#> • `` -> `...617`
#> • `` -> `...618`
#> • `` -> `...619`
#> • `` -> `...620`
#> • `` -> `...621`
#> • `` -> `...622`
#> • `` -> `...623`
#> • `` -> `...624`
#> • `` -> `...625`
#> • `` -> `...626`
#> • `` -> `...627`
#> • `` -> `...628`
#> • `` -> `...629`
#> • `` -> `...630`
#> • `` -> `...631`
#> • `` -> `...632`
#> • `` -> `...633`
#> • `` -> `...634`
#> • `` -> `...635`
#> • `` -> `...636`
#> • `` -> `...637`
#> • `` -> `...638`
#> • `` -> `...639`
#> • `` -> `...640`
#> • `` -> `...641`
#> • `` -> `...642`
#> • `` -> `...643`
#> • `` -> `...644`
#> • `` -> `...645`
#> • `` -> `...646`
#> • `` -> `...647`
#> • `` -> `...648`
#> • `` -> `...649`
#> • `` -> `...650`
#> • `` -> `...651`
#> • `` -> `...652`
#> • `` -> `...653`
#> • `` -> `...654`
#> • `` -> `...655`
#> • `` -> `...656`
#> • `` -> `...657`
#> • `` -> `...658`
#> • `` -> `...659`
#> • `` -> `...660`
#> • `` -> `...661`
#> • `` -> `...662`
#> • `` -> `...663`
#> • `` -> `...664`
#> • `` -> `...665`
#> • `` -> `...666`
#> • `` -> `...667`
#> • `` -> `...668`
#> • `` -> `...669`
#> • `` -> `...670`
#> • `` -> `...671`
#> • `` -> `...672`
#> • `` -> `...673`
#> • `` -> `...674`
#> • `` -> `...675`

StorageLocation <- file.path("C:", "Users", "JohnDoe", "Desktop")

SinglePlot <- "UV7-Gain"

TheSinglePlot <- QC_Plots(x = TidyData, MeasurementType = SinglePlot,
                         FailedFlag = TRUE, returntype="patchwork",
                         path=StorageLocation, filename="CytekAurora5L_QC")