Queries the available reference library for available fluorophores (and their naming conventions)

Queries the available reference library for available fluorophores (and their naming conventions)

Usage

QC_ReferenceLibrary(
  FluorNameContains,
  NumberDetectors,
  returnPlots = FALSE,
  plotlinecolor = NULL
)

Arguments

FluorNameContains

A character string pattern to match, example "APC"

NumberDetectors

The number of detectors the instrument contains. For Cytek Instruments 5L = 64, 4L_UV = 54, 4L_YG = 48, 3L=38, 2L_VB=30, 2L_BR=22, 1L=14 For BD S8=78, S6="48_S", A5="48_A" For Sony ID7000 7L=184, 6L_DUV="182_DUV", 5L=147, 4L=112, 3L=86 For ThermoFisher BigFoot 7L_488-561=55, 7L_532-594="52_7L", 6L_445="52_6L", 6L_785=51

returnPlots

Whether to return signature plot as well. Default FALSE.

plotlinecolor

Default NULL, otherwise if single line provide desired color

Value

A dataframe column containing matching Fluorophores from your querry

Examples

QC_ReferenceLibrary(FluorNameContains = "FITC", NumberDetectors=64)
#>       Fluorophore
#> 1            FITC
#> 2       FITC Plus
#> 3 Vio Bright FITC