Takes gated .fcs files and returns concentration and other info useful for Wetlab users.

Takes gated .fcs files and returns concentration and other info useful for Wetlab users.

Usage

Wetlab_Concentration(x, subset, nameKeyword, DilutionMultiplier, TotalVolume)

Arguments

x

A mapped gating set object

subset

Desired population node to derrive counts from

nameKeyword

Keyword containing samples name (ex. "GROUPNAME" or c("GROUPNAME", "TUBENAME"))

DilutionMultiplier

The dilution multiplier for the sample

TotalVolume

Volume the specimen was resuspended in.

Value

A data.frame of useful information

Examples

library(flowCore)
library(flowWorkspace)
library(CytoML)
library(dplyr)
library(purrr)

File_Location <- system.file("extdata", package = "Luciernaga")
WSP_File <- list.files(File_Location, pattern=".wsp", full.names = TRUE)
ws <- open_flowjo_xml(WSP_File[1])
gs <- flowjo_to_gatingset(ws, name=1, path = File_Location)
nameKeyword <- c("GROUPNAME", "TUBENAME")

TheData <- map(.x=gs, Wetlab_Concentration, subset = "CD45+",
  nameKeyword=nameKeyword, DilutionMultiplier=100, TotalVolume=1) %>%
   bind_rows()